Remove dead code: notifier module, unused fields, and minor cleanups
- Delete notifier.rs and notify-rust dependency (TTS replaced notifications)
- Remove max_direct_chars from TTSConfig (no longer used in speak pipeline)
- Remove unused VoiceGate::is_voice_idle() and wait_for_idle() methods
- Remove dead recording_config field from AudioRecorder
- Fix pointless match expression in hotkey.rs
- Remove redundant `as usize` cast in api.rs
- Remove unnecessary Copy-type rename in mcp_server.rs

+++ whisper-typer-rs/src/code_speaker/api.rs
@@ -411,7 +411,7 @@ async fn handle_user_input(
     );
 
     Json(SimpleResponse {
-        requeued: Some(requeued as usize),
+        requeued: Some(requeued),
         ..SimpleResponse::ok("ok")
     })
 }
